[TEMP] WIP for Ollama testing - will overwrite this commit
I've tested the OpenAI API in Insomnia, since that
is what Systems Group has explicitly said is supported.

I've also tested with Ollama because the OpenWeb UI
LLM API gateway they are using supports it, so I tried
to get lucky and see if the Systems Group's config supports
Ollama API.

Right now, I can't find the correct model name to use.
THe API endpoints that list models don't seem to work,
and guessing around with the capitalization etc. of the
model names from the OpenWeb UI user interface
does not result in finding the models.

I've reached out for help on the OpenAI endpoint and
also asked if we could use Ollama.